+++ App/scripts/common/UserService.js
@@ -32,6 +32,17 @@ sync.service('User', ['$http','$q','$rootScope',function Files ($http,$q,$rootSc
 		});
 		return promise.promise;
 	};
+	this.doSignUp = function(data) {
+		 var differed = $q.defer();
+      $http.post($rootScope.endPoint + '/register')
+      .success(function(response){
+        differed.resolve(response);
+      })
+      .error(function(err){
+        differed.reject(err);
+      })
+      return differed.promise;
+	};
 	this.groups = function(user){
       var differed = $q.defer();
       $http.get($rootScope.endPoint + '/api/v1/me/groups')
@@ -42,7 +53,7 @@ sync.service('User', ['$http','$q','$rootScope',function Files ($http,$q,$rootSc
         differed.reject(err);
       })
       return differed.promise;
-    }
+    };
 
 	return this;
 }])