+++ App/scripts/registerController.js
@@ -1,48 +1,39 @@
-Logger.controller('RegisterController', ['$scope','$rootScope','$http',function ($scope,$rootScope,$http) {
+var configuration = require('./configuration');
+
+sync.controller('RegisterController', ['$scope','$rootScope','$http',function ($scope,$rootScope,$http) {
     var options = {
         'password-notMatch': 'password do not match',
         'SignUpInProgress' : 'Wait we are setting up your account.'
     };
-    $scope.register=function(user){
+    $scope.doSignUp=function(user){
       $('.register-form-main-message').addClass('show success').html(options['SignUpInProgress']);
-        if($('#password').val() != $('#password-confirm').val()){
+        if(jQuery('#password').val() !== jQuery('#password-confirm').val()){
           $('.register-form-main-message').addClass('show error').html(options['password-notMatch']);
           setTimeout(messageRemove, 2000);
           function messageRemove(){
-              $('.register-form-main-message').removeClass('show error');
+              jQuery('.register-form-main-message').removeClass('show error');
           }
           return;
-        }
+        };
         var username=$('#username').val();
         var email=$('#email').val();
-
-
-        jQuery.post('/sessions', {username: username, password:user.password, email:email, option:user.option, phone:user.phone}, function(data, textStatus, xhr) {
-            if(data == 1){
+        jQuery.post($rootScope.endPoint+'/register', {username: username, password:user.password, email:email, option:'register', phone:user.phone}, function(data, textStatus, xhr) {
+            if(data.status === 200){
+                 if (!configuration.readSettings('user_credentials')) {
+                    configuration.saveSettings('user_credentials', [user.email, user.password]);
+                    //TODO on SignUp complete with success please use the above code to save user credential for future use also encrypt it on local!
+                };
                  Redirecting();
-            }else if(data ==0){
-                // console.log('we are fired this can not happen');
-            }
-        }).error(function(error) {
-
-            // console.log(error);
+            }else if(data.status ===500){
+                
+            };
         });
         function Redirecting(){
-            window.location = '/checkEmail';
-        }
-    }
-}]);
-
-Logger.directive('uniqueUsername', ['isUsernameAvailable',function(isUsernameAvailable) {
-    return {
-        restrict: 'A',
-        require: 'ngModel',
-        link: function(scope, element, attrs, ngModel) {
-            ngModel.$asyncValidators.uniqueUsername = isUsernameAvailable;
-        }
+            window.location = '#EmailConfirmation';
+        };
     };
 }]);
-Logger.factory('isUsernameAvailable', ['$q','$http','$rootScope',function($q, $http,$rootScope) {
+sync.factory('isUsernameAvailable', ['$q','$http','$rootScope',function($q, $http,$rootScope) {
     var options = {
         'btn-loading': '<i class="fa fa-spinner fa-pulse"></i>',
         'btn-success': '<i class="fa fa-check"></i>',
@@ -55,31 +46,29 @@ Logger.factory('isUsernameAvailable', ['$q','$http','$rootScope',function($q, $h
         'useAJAX': true,
     };
     return function(username) {
-
         var deferred = $q.defer();
-
         $http.get($rootScope.endPoint + '/api/v1/users?username=' + username + '&access_token=8EuqcMNkF2yP50Dicpv9hLRRp7WOSabPlCu22liY').success(function(data){
-            if(data=='available'){
-                $('.register-form-main-message').addClass('show success').html(options['msg-username-available']);
+            if(data === 'available'){
+                jQuery('.register-form-main-message').addClass('show success').html(options['msg-username-available']);
                 setTimeout(messageRemove, 2000);
                 function messageRemove(){
-                    $('.register-form-main-message').removeClass('show success');
+                    jQuery('.register-form-main-message').removeClass('show success');
                 }
-            }else if(data=='taken'){
-                $('.register-form-main-message').addClass('show error').html(options['msg-username-taken']);
+            }else if(data === 'taken'){
+                jQuery('.register-form-main-message').addClass('show error').html(options['msg-username-taken']);
                 setTimeout(usernameTaken, 2000);
                 function usernameTaken(){
-                    $('.register-form-main-message').removeClass('show error');
-                }
-            }
+                    jQuery('.register-form-main-message').removeClass('show error');
+                };
+            };
             deferred.reject();
         }).error(function(err) {
            deferred.resolve();
         });
         return deferred.promise;
-    }
+    };
 }]);
-Logger.directive('uniqueEmail', ['isEmailAvailable',function(isEmailAvailable) {
+sync.directive('uniqueEmail', ['isEmailAvailable',function(isEmailAvailable) {
     return {
         restrict: 'A',
         require: 'ngModel',
@@ -88,7 +77,16 @@ Logger.directive('uniqueEmail', ['isEmailAvailable',function(isEmailAvailable) {
         }
     };
 }]);
-Logger.factory('isEmailAvailable', ['$q','$http','$rootScope',function ($q, $http, $rootScope) {
+sync.directive('uniqueUsername', ['isUsernameAvailable',function(isUsernameAvailable) {
+    return {
+        restrict: 'A',
+        require: 'ngModel',
+        link: function(scope, element, attrs, ngModel) {
+            ngModel.$asyncValidators.uniqueUsername = isUsernameAvailable;
+        }
+    };
+}]);
+sync.factory('isEmailAvailable', ['$q','$http','$rootScope',function ($q, $http, $rootScope) {
     var options = {
         'btn-loading': '<i class="fa fa-spinner fa-pulse"></i>',
         'btn-success': '<i class="fa fa-check"></i>',
@@ -107,20 +105,20 @@ Logger.factory('isEmailAvailable', ['$q','$http','$rootScope',function ($q, $htt
 
         $http.get($rootScope.endPoint + '/api/v1/users?email=' + email + '&access_token=8EuqcMNkF2yP50Dicpv9hLRRp7WOSabPlCu22liY').success(function(data){
 
-            if(data=='email-available'){
-                $('.register-form-main-message').addClass('show success').html(options['msg-email-available']);
+            if(data==='email-available'){
+                jQuery('.register-form-main-message').addClass('show success').html(options['msg-email-available']);
                 setTimeout(messageRemove, 2000);
                 function messageRemove(){
-                    $('.register-form-main-message').removeClass('show success');
-                }
+                    jQuery('.register-form-main-message').removeClass('show success');
+                };
 
-            }else if(data=='email-taken'){
-                $('.register-form-main-message').addClass('show error').html(options['msg-email-taken']);
+            }else if(data==='email-taken'){
+                jQuery('.register-form-main-message').addClass('show error').html(options['msg-email-taken']);
                 setTimeout(messageEmailTaken, 2000);
                 function messageEmailTaken(){
-                    $('.register-form-main-message').removeClass('show error');
-                }
-            }
+                    jQuery('.register-form-main-message').removeClass('show error');
+                };
+            };
              deferred.reject();
          }).error(function() {
             deferred.resolve();