+++ __tests__/app/Auth/home.spec.js
@@ -1,14 +1,14 @@
 'use strict';
 
-describe('Controller: HomeController', function() {
+describe('Controller: RegisterController', function() {
     var rootScope, scope, controller, httpBackend;
 
-    beforeEach(module('Home'));
+    beforeEach(module('sync'));
 
     beforeEach(inject(function($controller, $rootScope, $httpBackend) {
         httpBackend = $httpBackend;
         scope = $rootScope.$new();
-        controller = $controller('HomeController', { '$scope': scope });
+        controller = $controller('RegisterController', { '$scope': scope });
         rootScope = $rootScope;
         scope.$digest();
     }));