+++ app_modules/configuration.js
@@ -13,7 +13,7 @@ function readSettings(settingKey) {
 }
 
 function getUserHome() {
-    return process.env[(process.platform == 'win32') ? 'USERPROFILE' : 'HOME'];
+    return process.env[(process.platform === 'win32') ? 'USERPROFILE' : 'HOME'];
 }
 
 module.exports = {