+++ main.js
@@ -4,25 +4,20 @@ var app = require('app');
 var BrowserWindow = require('browser-window');
 var path = require('path');
 var mainWindow = null;
-
 var globalShortcut = require('global-shortcut');
 var _ = require('lodash');
 var fs = require('file-system');
-
 var shell = require('shelljs');
-
-var notifier = require('node-notifier');
+var notification = require('./app_modules/notifier');
 var child_process = require('child_process');
 var graphql = require ('graphql').graphql;
 var chokidar= require('chokidar');
- var mkdirp = require('mkdirp');
+var mkdirp = require('mkdirp');
 var filessystem = require('fs');
 var ncp = require('ncp').ncp;
 var express = require('express'), cors = require('cors');
-
 var http = require('http');
-
-var configuration = require('./configuration');
+var configuration = require('./app_modules/configuration');
 var options = {
 	"debug": true,
 	"version": "1.0.0",
@@ -30,17 +25,14 @@ var options = {
 	"root_view": "index.html"
 };
 require('crash-reporter').start();
-
-
 app.on('ready', function() {
     if (!configuration.readSettings('shortcutKeys')) {
         configuration.saveSettings('shortcutKeys', ['ctrl', 'shift']);
-    }
-    
+        //TODO on SignUp complete with success please use the above code to save user credential for future use also encrypt it on local!
+    };
     mainWindow = new BrowserWindow({
         frame: true,
         kiosk: true,
-        
         height: 500,   
         title:"StreamUpBox Setup",
         resizable: false,
@@ -49,26 +41,26 @@ app.on('ready', function() {
     mainWindow.setMenu(null);
     mainWindow.loadUrl(path.join('file://', __dirname, options.views_dir, options.root_view));
 
-    // setGlobalShortcuts();
-
+    
     var folderWatcher = function(object) {
        
         require('chokidar').watch('/home/StreamUpBox', {ignored: /[\/\\]\./}).on('all', function(event, path) {
             if(event === "unlink"){
-                notie('removed files');
+                notification.send('removed files');
             }else if(event === "add"){
                 //we got a new files push them to the server
+                //TODO As we are listening on new files added please upload each on server
                  var server = express();
                     var corsOptions = {
                     origin: 'http://localhost:8000/'
                     };
-                    server.post('/api/v1/upload', cors(corsOptions), function(req, res, next){
+                    server.post('streamupbox.com/api/v1/upload', cors(corsOptions), function(req, res, next){
                         console.log(res);
                         res.json({msg: 'This is CORS-enabled for only example.com.'});
                     });
-                 notie('add new file');
+                 notification.send('add new file');
             }else if(event ==="change"){
-                 notie('changes');
+                 notification.send('changes');
 
             }
             
@@ -76,7 +68,7 @@ app.on('ready', function() {
         });
     };
     
-    //Creating dir
+ 
     var dir = '/home/StreamUpBox';
     if (!filessystem.existsSync(dir)){
         filessystem.mkdirSync(dir);
@@ -92,19 +84,19 @@ app.on('ready', function() {
     });
 });
 
-// function setGlobalShortcuts() {
-//     globalShortcut.unregisterAll();
+function setGlobalShortcuts() {
+    globalShortcut.unregisterAll();
 
-//     var shortcutKeysSetting = configuration.readSettings('shortcutKeys');
-//     var shortcutPrefix = shortcutKeysSetting.length === 0 ? '' : shortcutKeysSetting.join('+') + '+';
+    var shortcutKeysSetting = configuration.readSettings('shortcutKeys');
+    var shortcutPrefix = shortcutKeysSetting.length === 0 ? '' : shortcutKeysSetting.join('+') + '+';
 
-//     globalShortcut.register(shortcutPrefix + '1', function () {
-//         mainWindow.webContents.send('global-shortcut', 0);
-//     });
-//     globalShortcut.register(shortcutPrefix + '2', function () {
-//         mainWindow.webContents.send('global-shortcut', 1);
-//     });
-// }
+    globalShortcut.register(shortcutPrefix + '1', function () {
+        mainWindow.webContents.send('global-shortcut', 0);
+    });
+    globalShortcut.register(shortcutPrefix + '2', function () {
+        mainWindow.webContents.send('global-shortcut', 1);
+    });
+};
 
 var ipc = require('ipc');